test: operators cleanup

+++ upgrades/pre/upgrade-2-13-0.sh
@@ -2,9 +2,17 @@
 
 set -eu
 
-kubectl delete job gitea-prepare -n maintenance
-kubectl delete job harbor -n maintenance
-kubectl delete job keycloak -n maintenance
+if [[ $(kubectl get job gitea-prepare -n maintenance 2>/dev/null) ]]; then
+  kubectl delete job gitea-prepare -n maintenance
+fi
+
+if [[ $(kubectl get job harbor -n maintenance 2>/dev/null) ]]; then
+  kubectl delete job harbor -n maintenance
+fi
+
+if [[ $(kubectl get job keycloak -n maintenance 2>/dev/null) ]]; then
+  kubectl delete job keycloak -n maintenance
+fi
 
 if [[ $(helm status -n gitea-operator gitea-operator 2>/dev/null) ]]; then
   helm uninstall -n gitea-operator gitea-operator
